requirements.txt
fastapi
uvicorn

python-dotenv
sqlalchemy
pydantic

python-jose
pymysql

python-jose[cryptography]
 passlib==1.7.4
python-multipart
email_validator
bcrypt==3.2.0
pydantic-settings

# Translation
deep-translator

# Text to Speech
gTTS

google-genai
ollama
googletrans 

# OCR core
pytesseract
pillow
opencv-python
pdf2image

# Utilities
numpy
